requirements.txt
appdirs==1.4.3
configparser==3.5.0
Django==2.0.1
django-graphos==0.3.41
django-user-agents==0.3.0
djangorestframework==3.7.7
django-settings-export=1.2.1
et-xmlfile==1.0.1
jdcal==1.3
lml==0.0.1
pymysql==0.8.0
packaging==16.8
pyparsing==2.2.0
pytz==2016.10
requests==2.13.0
six==1.10.0
texttable==1.1.1
ua-parser==0.7.3
user-agents==1.1.0
XlsxWriter==1.0.2